test_client/requirements.txt
pytest==8.1.1
allure-pytest==2.13.3
allure-python-commons==2.13.3
Appium-Python-Client==3.2.1
